updating personal finance calculator

+++ python_introduction/finance_calculator.py
@@ -3,7 +3,7 @@ monthly_income= input("Enter your monthly income: ")
 
 monthly_expenses= input("Enter your monhtly expenses: ")
 
-monthly_savings = (int(monthly_income) - int(monthly_expenses))
+monthly_savings = float(monthly_income) - float(monthly_expenses)
 
 projected_savings = monthly_savings * 12 + (monthly_savings * 12 * 0.05)
